+++ smacc2_client_library/nav2z_client/custom_planners/backward_local_planner/src/backward_local_planner/backward_local_planner.cpp
@@ -735,7 +735,7 @@ geometry_msgs::msg::TwistStamped BackwardLocalPlanner::computeVelocityCommands(
   // check plan rejection
   bool acceptedLocalTrajectoryFreeOfObstacles = true;
 
-  uint64_t mx, my;
+  unsigned int mx, my;
 
   if (this->enable_obstacle_checking_)
   {
